refactor: split code and tests
