+++ package.json
@@ -11,7 +11,7 @@
   "scripts": {
     "doctor": "yarn dlx @yarnpkg/doctor .",
     "exec:eslint": "eslint --cache",
-    "exec:jest": "jest --coverage --color --forceExit",
+    "exec:jest": "jest --coverage --color --detectOpenHandles",
     "exec:prettier": "prettier  --config .prettierrc --ignore-path .prettierignore --no-editorconfig",
     "exec:foreach": "yarn workspaces foreach -vpit --exclude kalos",
     "lint:markdown": "markdownlint -c markdownlint.yml $(find . -name '*.md' -not -path './.yarn/*')",
